model/requirements.txt
pandas
groq
python-dotenv
selenium
beautifulsoup4
requests
langgraph